integrinMediatedCellAdhesion {graph}	R Documentation

KEGG Integrin Mediated Cell Adhesion graph

Description

A graph representing the integrin-mediated cell adhesion pathway from KEGG, as well as a list of attributes for use in plotting the graph with Rgraphviz.

Usage


data(integrinMediatedCellAdhesion)

Details

The integrinMediatedCellAdhesion data set contains two objects:

The first is IMCAGraph, which is an object of class graph-NEL and represents the hsa04510 graph from KEGG.

The second is IMCAAttrs, which is a list of four elements. The first element, defAttrs corresponds to the attrs arguments of agopen and plot.graph. The second element is nodeAttrs which corresponds to the nodeAttrs argument in the same two functions from Rgraphviz. The third element, subGList corresponds to the subGList argument in those functions. Lastly, the fourth element, LocusLink provides a named list where the names are the nodes and the values are vectors of LocusLink ID values which correspond to those nodes.

The values from defAttrs, nodeAttrs and subGList in the IMCAAttrs list are part of an ongoing attempt by Bioconductor to provide the set of options to most accurately recreate the actual visual image of the pathway from the KEGG site using Rgraphviz. Users may try out their own combination of attributes and settings for their own needs, but these represent our own efforts at as closely recreating the image as possible.

Source

http://www.genome.ad.jp/kegg/pathway/hsa/hsa04510.html

Examples


data(integrinMediatedCellAdhesion)
if (require("Rgraphviz") & interactive())
  plot(IMCAGraph, attrs=IMCAAttrs$defAttrs,
       nodeAttrs=IMCAAttrs$nodeAttrs, subGList=IMCAAttrs$subGList)

[Package graph version 1.60.0 Index]