MAPKsig {graph}	R Documentation

A graph encoding parts of the MAPK signaling pathway

Description

A graph encoding parts of the MAPK signaling pathway

Usage

data(MAPKsig)

Format

The format is: Formal class 'graphNEL' [package "graph"] with edgemode "directed".

Source

The KEGG pancreatic cancer pathway was visually inspected on 17 Sept 2007, and the subgraph associated with MAPK signaling was isolated. The HUGO names for each symbol in the KEGG visualization were obtained and checked for existance on hgu95av2. Some abbreviated terms for processes are also included as nodes.

Examples


data(MAPKsig)
if (require(Rgraphviz)) {
   nat = rep(FALSE, length(nodes(MAPKsig)))
   names(nat) = nodes(MAPKsig)
   plot(MAPKsig, nodeAttrs=list(fixedsize=nat))
   }

[Package graph version 1.60.0 Index]