adj-methods {graph}	R Documentation

Methods for finding the adjacency list for selected nodes.

Description

This generic function takes an object that inherits from the graph class and a node in that graph and returns a vector containing information about all other nodes that are adjacent to the given node. This means that they are joined to the given node by an edge. The accessibility list, acc is the list of all nodes that can be reached from a specified node.

Value

The methods return vector of nodes that are adjacent to the specified node.

Methods

object = graph

An object that inherits from glass graph

index

An index (could be multiple) which can be either the integer offset for the node(s) or their labels.

See Also

acc-methods

Examples


 set.seed(123)
 gR3 <- randomGraph(LETTERS[1:4], M<-1:2, p=.5)
 adj(gR3, "A")
 adj(gR3, c(2,3))

[Package graph version 1.60.0 Index]