apoptosisGraph {graph}	R Documentation

KEGG apoptosis pathway graph

Description

A graph representing the apoptosis pathway from KEGG, as well as a data.frame of attributes for use in plotting the graph with Rgraphviz and a list to compare the nodes with their respective LocusLink IDs.

Usage


data(apopGraph)

Details

The apopGraph data set contains three objects:

The first is apopGraph, which is an object of class graph-NEL and represents the hsa04210 graph from KEGG.

The second is apopAttrs, which is a data.frame with two columns, and a row for every node in apopGraph. The first column lists what color the node is represented with on the KEGG site. The second column lists the type of the node - either genesym or text. Most nodes are of type genesym as they represent genes, but some of the nodes in the KEGG graph were not genes and thus those nodes are of type text.

The third, apopLocusLink is a named list where the names correspond to the node names in apopGraph. The values of the list are the LocusLink IDs that correspond to that node in the KEGG graph.

Source

http://www.genome.ad.jp/kegg/pathway/hsa/hsa04210.html

Examples


data(apopGraph)
if (require("Rgraphviz") & interactive())
  plot(apopGraph)

[Package graph version 1.60.0 Index]