attrDefaults-methods {graph}	R Documentation

Get and set the default attributes of an attrData object

Description

The attrDefaults method provides access to a attrData-class object's default attribute list. The default attribute list of a attrData-class object defines what attributes can be customized for individual data elements by defining attribute names and default values.

Usage


attrDefaults(self, attr)
attrDefaults(self, attr) <- value

Arguments

self
A attrData-class instance

attr
A character vector of length 1 giving the name of an attribute. Can be missing.

value
An R object that will be used as the default value of the specified attribute, or a named list of attribute name/default value pairs if attr is missing.

[Package graph version 1.60.0 Index]