aveNumEdges {graph}	R Documentation

Calculate the average number of edges in a graph

Description

aveNumEdges divides the number of edges in the graph by the number of nodes to give the average number of edges.

Usage


aveNumEdges(objgraph)

Arguments

objgraph
the graph object

Value

A double representing the average number of edges will be returned.

Author(s)

Elizabeth Whalen

See Also

numEdges, mostEdges, numNoEdges

Examples


 set.seed(124)
 g1 <- randomGraph(1:10, letters[7:12], p=.6)
 aveNumEdges(g1)


[Package graph version 1.60.0 Index]