biocRepos {graph}	R Documentation

A graph representing the Bioconductor package repository

Description

This graph is a rendition of the Bioconductor package repository and represents the dependency graph of that repository. An edge between two package denotes a dependency on the 'to' package by the 'from' package.

Usage

data(biocRepos)

Examples


   data(biocRepos)
   ## An example of usage will be here soon

[Package graph version 1.60.0 Index]