calcProb {graph}	R Documentation

Calculate the hypergeometric probability of the subgraph's number of edges.

Description

calcProb calculates the probability of having the number of edges found in the subgraph given that it was made from origgraph. The hypergeometric distribution is used to calculate the probability (using the pdf).

Usage


calcProb(subgraph, origgraph)

Arguments

subgraph
subgraph made from the original graph

origgraph
original graph object from which the subgraph was made

Value

The probability of the subgraph's number of edges is returned.

Author(s)

Elizabeth Whalen

See Also

calcSumProb

Examples


 #none right now

[Package graph version 1.60.0 Index]