clearNode {graph}	R Documentation

clearNode

Description

This function removes all edges to or from the specified node in the graph.

Usage


clearNode(node, object)

Arguments

node
a node

object
a graph

Details

All edges to and from node are removed. node can be a vector.

Value

A new instance of the graph with all edges to and from the specified node(s) removed.

Author(s)

R. Gentleman

See Also

removeNode, removeEdge

Examples


 V <- LETTERS[1:4]
 edL3 <- vector("list", length=4)
 for(i in 1:4)
   edL3[[i]] <- list(edges=(i%%4)+1, weights=i)
 names(edL3) <- V
 gR3 <- graphNEL(nodes=V, edgeL=edL3, "directed")
 g4 <- clearNode("A", gR3)


[Package graph version 1.60.0 Index]