duplicatedEdges {graph}	R Documentation

duplicatedEdges

Description

A multigraph is a graph where edges between nodes can be represented several times. For some algorithms this causes problems. duplicatedEdges tests an instance of the graphNEL class to see if it has duplicated edges and returns TRUE if it does and FALSE otherwise.

Usage


duplicatedEdges(graph)

Arguments

graph
An instance of the class graphNEL

Details

It would be nice to handle other types of graphs.

Value

A logical, either TRUE if the graph has duplicated edges or FALSE it not.

Author(s)

R. Gentleman

See Also

connComp, ugraph

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,

[Package graph version 1.60.0 Index]