edgeData-methods {graph}	R Documentation

Get and set attributes for the edges of a graph object

Description

Attributes of the edges of a graph can be accessed using edgeData. The attributes must be defined using edgeDataDefaults. You can ommit the from or to argument to retrieve attribute values for all edges to (respectively, from) a given node.

Usage


edgeData(self, from, to, attr)
edgeData(self, from, to, attr) <- value

Arguments

self
A graph-class instance

from
A character vector of node names

to
A character vector of node names

attr
A character vector of length one specifying the name of a node attribute

value
An R object to store as the attribute value

[Package graph version 1.60.0 Index]