edgeDataDefaults-methods {graph}	R Documentation

Get and set default attributes for the edges of a graph

Description

Set default values for attributes associated with the edges of a graph.

Usage


edgeDataDefaults(self, attr)
edgeDataDefaults(self, attr) <- value

Arguments

self
A graph-class instance

attr
A character vector of length one giving the name of the attribute

value
An R class to use as the default value for the specified attribute

[Package graph version 1.60.0 Index]