edgeSets {graph}	R Documentation

MultiGraph edgeSet data

Description

C57BL/6J and C3H/HeJ mouse strains exhibit different cardiovascular and metabolic phenotypes on the hyperlipidemic apolipoprotein E (Apoe) null background. The interaction data for the genes from adipose, brain, liver and muscle tissue samples from male and female mice are included as a list of data.frames. Each data.frame contains information for the from-gene, to-gene and the strength of interaction (weight) for each of the tissues studied.

Usage


    data(esetsFemale)
    data(esetsMale)

Source

Sage Commons Repository http://sagebase.org/commons/dataset1.php#UCLA1

Examples


data(esetsFemale)
data(esetsMale)


[Package graph version 1.60.0 Index]