graphExamples {graph}	R Documentation

A List Of Example Graphs

Description

This data set contains a list of example graphNEL objects, which can then be used for plotting.

Usage

data(graphExamples)

Source

Various sources, primarily from randomGraph and randomEGraph

Examples


data(graphExamples)
a <- graphExamples[[1]]
a

[Package graph version 1.60.0 Index]