inEdges {graph}	R Documentation

Generic Method inEdges

Description

Returns a list of all incoming edges for the specified nodes.

Usage


inEdges(node, object)

Arguments

node
character vector of node names

object
a graph object

Details

If no node argument is specified, inEdges returns the incoming edges for all nodes in the graph.

For an undirected graph, inEdges returns all edges for the specified nodes.

Value

A list with length matching the length of node. If node was missing, a list containing an element for each node in the graph.

Each list element contains a character vector of node names giving the nodes that have outgoing edges to the node given by the name of the list element.

Author(s)

R. Gentleman

See Also

removeNode, clearNode

Examples



  V <- LETTERS[1:4]
  edL3 <- vector("list", length=4)
  for(i in 1:4)
    edL3[[i]] <- list(edges=(i%%4)+1, weights=i)
  names(edL3) <- V
  gR3 <- graphNEL(nodes=V, edgeL=edL3, "directed")
  inEdges(c("A", "B"), gR3)


[Package graph version 1.60.0 Index]