isAdjacent-methods {graph}	R Documentation

Determine if nodes share an edge in a graph

Description

For a given subclass of graph-class, returns TRUE if the graph contains an edge from node specified by from to the node specified by to.

The appropriate logical vector will be returned as long as from and to have the same length and contain nodes in the graph object specified by object.

Usage


isAdjacent(object, from, to, ...)

Arguments

object
An instance of a subclass of graph-class.

from
A character vector of nodes in the graph.

to
A character vector of nodes in the graph

...
May be used by methods called on subclasses of graph

[Package graph version 1.60.0 Index]