mostEdges {graph}	R Documentation

Find the node in a graph with the greatest number of edges

Description

mostEdges finds the node that has the most edges in the graph. This is the node with the highest degree.

Usage


mostEdges(objGraph)

Arguments

objGraph
the graph object

Value

index
the index of the node with the most edges

id
the node value with the most edges; may be affy id, locus link id, or genename depending on the node type

maxLen
the number of edges for that node

Author(s)

Elizabeth Whalen

See Also

numEdges, aveNumEdges, numNoEdges

Examples


  set.seed(123)
  g1 <- randomGraph(11:30, letters[20:26], p=.4)
  mostEdges(g1)

[Package graph version 1.60.0 Index]