multiGraph-class {graph}	R Documentation

Class "multiGraph"

Description

A collection of classes to model multigraphs. These include the multiGraph class as well as classes to contain edge sets.

Objects from the Class

Objects can be created from the multiGraph class, the edgeSet class is virtual, and particular variants should be used.

Slots

These slots are for the multiGraph class.

nodes

The names of the nodes.

edgeL

A list of edge lists.

nodeData

An instance of the attrData class.

graphData

A list.

These slots are for the edgeSet class, or one of its sublcasses.

edgeData

An instance of the attrData class.

edgemode

A character vector, one of directed, or undirected.

edgeL

A list of the edges (graphNEL)

adjMat

An adjacency matrix (graphAM)

Methods

show

Print a multigraph.

isDirected

A vector indicating which of the edgeSets is directed.

nodes

Retrieve the node names

numNodes

Return the number of nodes

edges

Return either all edges, or a subset of them, depending on the arguments supplied.

numEdges

Return a vector with the number of edges, for each edge set.

[Package graph version 1.60.0 Index]