nodeData-methods {graph}	R Documentation

Get and set attributes for the nodes of a graph object

Description

Attributes of the nodes of a graph can be accessed using nodeData. The attributes must be defined using nodeDataDefaults. You can ommit the n argument to retrieve attributes for all nodes in the graph. You can ommit the attr argument to retrieve all attributes.

Usage


nodeData(self, n, attr)
nodeData(self, n, attr) <- value

Arguments

self
A graph-class instance

n
A character vector of node names

attr
A character vector of length one specifying the name of a node attribute

value
An R object to store as the attribute value

[Package graph version 1.60.0 Index]