nodeDataDefaults-methods {graph}	R Documentation

Get and set default attributes for the nodes of a graph

Description

You can associate arbitrary attributes with the nodes of a graph. Use nodeDataDefaults to specify the set of attributes that describe nodes. Each attribute must have a default value. You can set the attribute for a particular node or set of nodes using nodeData.

Usage


nodeDataDefaults(self, attr)
nodeDataDefaults(self, attr) <- value

Arguments

self
A graph-class instance

attr
A character vector of length one giving the name of an attribute

value
An R object to set as the default value for the given attribute

[Package graph version 1.60.0 Index]