numNoEdges {graph}	R Documentation

Calculate the number of nodes that have an edge list of NULL

Description

numNoEdges calculates the number of nodes that have an edge list of NULL (i.e. no edges).

Usage


numNoEdges(objGraph)

Arguments

objGraph
the graph object

Value

An integer representing the number of NULL edge lists in the graph.

Author(s)

Elizabeth Whalen

See Also

numEdges, aveNumEdges, mostEdges

Examples


  set.seed(999)
  g1 <- randomEGraph(letters, .01)
  numNoEdges(g1)

[Package graph version 1.60.0 Index]