randomEGraph {graph}	R Documentation

Random Edge Graph

Description

A function to create random graphs according to a random edge model. The user supplies the set of nodes for the graph as V and either a probability, p, that is used for each edge or the number of edges, edges they want to have in the resulting graph.

Usage


randomEGraph(V, p, edges)

Arguments

V
The nodes for the graph.

p
The probability of an edge being selected.

edges
The number of edges wanted.

Details

The user must specify the set of nodes and either a probability for edge selection or the number of edges wanted, but not both. Let nV denote the number of nodes. There are choose(nV, 2) edges in the complete graph. If p is specified then a biased coin (probability of heads being p) is tossed for each edge and if it is heads that edge is selected. If edges is specified then that many edges are sampled without replacement from the set of possible edges.

Value

An object of class graphNEL-class that contains the nodes and edges.

Author(s)

R. Gentleman

See Also

randomGraph

Examples


  set.seed(123)
  V <- letters[14:22]
  g1 <- randomEGraph(V, .2)

  g2 <- randomEGraph(V, edges=30)

[Package graph version 1.60.0 Index]