graph.par {graph}	R Documentation

Graphical parameters and other settings

Description

Functions providing an interface to persistent graphical parameters and other settings used in the package.

Usage


graph.par(...)
graph.par.get(name)

Arguments

...
either character strings naming parameters whose values are to be retrieved, or named arguments giving values that are to be set.

name
character string, giving a valid parameter name.

Details

graph.par works sort of like par, but the details are yet to be decided.

graph.par.get(name) is equivalent to graph.par(name)[[1]]

Value

In query mode, when no parameters are being set, graph.par returns a list containing the current values of the requested parameters. When called with no arguments, it returns a list with all parameters. When a parameter is set, the return value is a list containing previous values of these parameters.

Author(s)

Deepayan Sarkar, deepayan.sarkar@r-project.org

See Also

par

[Package graph version 1.60.0 Index]