simpleEdge-class {graph}	R Documentation

Class "simpleEdge".

Description

A simple class for representing edges in graphs.

Objects from the Class

Objects can be created by calls of the form new("simpleEdge", ...).

Slots

edgeType:

Object of class "character"; the type of the edge.

weight:

Object of class "numeric"; the edge weight.

directed:

Object of class "logical"; is the edge directed.

bNode:

Object of class "character"; the beginning node.

eNode:

Object of class "character"; the ending node.

Methods

No methods defined with class "simpleEdge" in the signature.

Note

All slots are length one vectors (this is not currently checked for). If the edge is not directed there is no real meaning to the concepts of beginning node or ending node and these should not be interpreted as such.

Author(s)

R. Gentleman

Examples


  new("simpleEdge", bNode="A", eNode="D")

[Package graph version 1.60.0 Index]