subGraph {graph}	R Documentation

Create a Subgraph

Description

Given a set of nodes and a graph this function creates and returns subgraph with only the supplied nodes and any edges between them.

Usage


subGraph(snodes, graph)

Arguments

snodes
A character vector of node labels.

graph
A graph object, it must inherit from the graph class.

Details

The returned subgraph is a copy of the graph. Implementations for Implementations for graphNEL, distGraph and clusterGraph.

Value

A graph of the same class as the graph argument but with only the supplied nodes.

Author(s)

R. Gentleman

See Also

nodes,edges

Examples


        set.seed(123)
        x <- rnorm(26)
        names(x) <- letters
        library(stats)
        d1 <- dist(x)
        g1 <- new("distGraph", Dist=d1)
        subGraph(letters[1:5], g1)


[Package graph version 1.60.0 Index]