toDotR-methods {graph}	R Documentation

Methods for Function toDotR, using R to generate a dot serialization

Description

There are two basic methods of generating dot (http://www.graphviz.org) language serializations of R graph-class structures. First, using the toDot methods of the Rgraphviz package, the native graphviz agraph-associated methods can be employed to create the dot serialization. Second, with the methods described here, R functions can be used to perform the serialization directly from the graph data structure, without Rgraphviz.

Methods

G = "graphNEL", outDotFile = "character", renderList = "list", optList = "list"

create dot language descriptionof graph

G = "graphNEL", outDotFile = "character", renderList = "missing", optList = "missing"

create dot language descriptionof graph

G = "graphNEL", outDotFile = "character", renderList = "missing", optList = "list"

create dot language descriptionof graph

G = "graphNEL", outDotFile = "missing", renderList = "missing", optList = "missing"

create dot language descriptionof graph

G = "graphNEL", outDotFile = "missing", renderList = "missing", optList = "list"

create dot language descriptionof graph

G = "graphNEL", outDotFile = "missing", renderList = "character", optList = "missing"

create dot language descriptionof graph

G = "graphNEL", outDotFile = "missing", renderList = "list", optList = "list"

create dot language descriptionof graph

G = "graphNEL", outDotFile = "missing", renderList = "list", optList = "missing"

create dot language descriptionof graph

G = "compoundGraph", outDotFile = "character", renderList = "list", optList = "missing"

create dot language descriptionof graph

G = "compoundGraph", outDotFile = "character", renderList = "list", optList = "list"

create dot language descriptionof graph

G = "compoundGraph", outDotFile = "missing", renderList = "list", optList = "missing"

create dot language descriptionof graph

See Also

toDot-methods

Examples


example(randomGraph)
tmp <- tempfile()
toDotR( g1, tmp )
readLines(tmp)
unlink(tmp)

[Package graph version 1.60.0 Index]