validGraph {graph}	R Documentation

Test whether graph object is valid

Description

validGraph is a validating function for a graph object.

Usage


validGraph(object, quietly=FALSE)

Arguments

object
a graph object to be tested

quietly
TRUE or FALSE indicating whether output should be printed.

Value

If the graph object is valid, TRUE is returned otherwise FALSE is returned. If object is not a valid graph and quietly is set to FALSE then descriptions of the problems are printed.

Author(s)

Elizabeth Whalen

See Also

graph-class

Examples


  
  testGraph<-graphNEL()
  testGraph@nodes<-c("node1","node2","node3")
  validGraph(testGraph)


[Package graph version 1.60.0 Index]