write.tlp {graph}	R Documentation

Write a graph object in a file in the Tulip format

Description

Write a graph object in a file in the Tulip format.

Usage


write.tlp(graph, filename)

Arguments

graph
a graph object

filename
Name of the output file

Details

The Tulip format is used by the program Tulip.

Author(s)

Laurent Gautier <laurent@cbs.dtu.dk>

References

http://www.tulip-software.org/

[Package graph version 1.60.0 Index]